refactor: extract Correlation helper class
